Testing scene change

+++ Slice/SliceEngine/src/ECS/GOFactory.h
@@ -167,7 +167,7 @@ namespace SliceEngine
 		void Unparent(Entity entity);
 		void SetParent(Entity entity, Entity parentEntity = entt::null);
 		void SetNewSceneGraphLocation(Entity targetEntity, Entity leftEntity, Entity rightEntity);
-		void BuildSceneGraph(std::unordered_map<uint64_t, uint64_t> map);
+		void BuildSceneGraph(std::unordered_map<uint32_t, uint32_t> map);
 		void ClearGameObjects();
 		void UpdateTransformFromParent(Entity entity, Entity parent);
